+++ Examples/UIKit-Swift/Superwall-UIKit-Swift/TrackEventViewController.swift
@@ -50,7 +50,7 @@ final class TrackEventViewController: UIViewController {
   @IBAction private func logOut() {
     UserDefaults.standard.setValue(false, forKey: "IsLoggedIn")
     Task {
-      await SuperwallService.logOut()
+      await SuperwallService.reset()
       _ = navigationController?.popToRootViewController(animated: true)
     }
   }